Fix:Update Square_root.py

+++ Calculator/Square_root.py
@@ -1,3 +1,3 @@
 def square_root(a):
     a = int(a)
-    return round(a, 9)
+    return float(round(a, 9))
